refactor(app structure): continued extracting various intents/actions to the correct models folders
